Remove randr from utility.jl
