+++ src/samplers/pgibbs.jl
@@ -140,7 +140,7 @@ assume{T<:Union{PG,SMC}}(spl::Sampler{T}, dist::Distribution, vn::VarName, _::Va
     if ~haskey(vi, vn)
       r = rand(dist)
       push!(vi, vn, r, dist, spl.alg.gid)
-      spl.info[:cache_updated] = 0b00   # sanity flag mask for getidcs and getranges
+      spl.info[:cache_updated] = CACHERESET # sanity flag mask for getidcs and getranges
       r
     elseif isnan(vi, vn)
       r = rand(dist)
